refactor: consolidate logic, prevent unnecessary API calls
The initial version of this was making a duplicate call for the first page of categories. This refactors it so it uses the initial call to both grab the count and the first page of results.

I also extracted this logic and put it in a shared lib function.
